requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
kafka-python